==> R/00_all.qmd <==
---
title: "00_all"
author: Lucía de Lamadrid, Eric Torres, Konstantina Gkopi, Elena Iriondo and Jorge Santiago
format:
  html:
    embed-resources: true
---

### 1. Load packages

```{r}
library("tidyverse")
library("quarto")
library("fs")
library("here")
```

### 2. Render Quarto documents

We are going to find all the Quarto documents in the R folder (i.e, those that end in .qmd, with the exception of 00_all.qmc)

```{r}
source_dir <- here("R")

# List all .qmd files in the directory, with the exception of 00_all.qmd
qmd_files <- dir_ls(source_dir, glob = "*.qmd") |> 
  discard(~ basename(.x) == "00_all.qmd")  

cat("The following files will be rendered:\n")
print(qmd_files)

# Render each file
qmd_files |> 
  walk(~ quarto_render(.x))

cat("Rendering complete.\n")
```

### 3. Move html files to result folder and delete unwanted folders

Then, we will move the .html files to the results folder. In the source directory, some folders (whose name end with "\_files") are also created, which include the figures of the plots in their respective Quarto document. We don't want to keep them, because these figures are already being saved in the results folder when executing the .qmd files.

```{r}
destination_dir <- here("results")

# Create the destination folder in case it does not exist
dir_create(destination_dir)


rendered_files <- dir_ls(source_dir, glob = "*.html")
cat("Moving the following rendered files to the results folder:\n")
print(rendered_files)

# Move the rendered files (.html) to the results directory
rendered_files |> 
  walk(~ file_move(.x, path(destination_dir, basename(.x))))


# Remove '_files' folders
folders_to_remove <- dir_ls(source_dir, type = "directory") |> 
  keep(~ str_detect(basename(.x), "_files$"))
cat("Removing the following '_files' folders:\n")
print(folders_to_remove)

folders_to_remove |> 
  walk(dir_delete)

cat("File move and folder cleanup complete.\n")
```

### Rendering the presentation 

```{r}
quarto::quarto_render(here("doc/presentation.qmd"))
```
