==> R/01_load.qmd <==
---
title: "Data loading"
author: Eric Torres, Lucía de Lamadrid, Konstantina Gkopi, Elena Iriondo, Jorge Santiago
format:
  html:
    embed-resources: true
editor: visual
---

### 1. Load packages

```{r}
#| message: false 
rm(list=ls())
library("tidyverse") 
library("googledrive")
library("readr")
```

### 2. Download and read files

With the following code chunk, data is accessed programmatically from an ID in Google Drive. The user has to authenticate defined with their Google account in order to retrieve the datasets:

```{r}
#| message: FALSE 

# Definition of the folder in which we will store the datasets
raw_folder <- "../_raw"

# Check if the folder exists; create it if not
if (!dir.exists(raw_folder)) {
  dir.create(raw_folder)
  message("Created folder: ", raw_folder)
} else {
  message("Folder already exists: ", raw_folder)
}

# File ID on Google Drive (more direct approach)
file_id <- "1CW63f5yjQsfovsbEJECvi1NlxkAIEer4"

# Definition of the destination path 
dest_path <- file.path(raw_folder, "Microbiome_data.zip")

# Download the file into _raw
drive_download(as_id(file_id), 
               path = dest_path, 
               overwrite = TRUE)

# Unzip 
unzip(dest_path, 
      exdir = raw_folder)
```

We import three datasets:

-   MicrobiomeWithMetadata.csv: it contains the relative abundance of the OTUs analysed per observation, as well as the conditions of each sample.

-   MicrobiomeMetadataDictionary.csv: it has the equivalence between the numerical values and the conditions.

-   MicrobiomeOTUtaxonomy.csv: it contains the taxonomic ranks of each analyses OTU.

```{r}
data_file_metadata <- "/data_zip/MicrobiomeWithMetadata.csv" 
data_file_dictionary <- "/data_zip/MicrobiomeMetadataDictionary.csv" 
data_file_otu <- "/data_zip/MicrobiomeOTUtaxonomy.csv"  

# Definition of the pathways of the datasets.
data_path_metadata = str_c(raw_folder, data_file_metadata) 
data_path_dictionary = str_c(raw_folder, data_file_dictionary) 
data_path_otu = str_c(raw_folder, data_file_otu) 
```

```{r}
#| message: false  
# Load data 
metadata_df <- read_csv(data_path_metadata) 
dictionary_df <- read_csv(data_path_dictionary)
otu_df <- read_csv(data_path_otu)  
```

In the `otu_df` there are two types of separators (`\t` followed by `,`) in the first column, so we will eliminate the `\t` to have a uniform data frame.

```{r}
otu_df <- otu_df |> 
  mutate(across(everything(), 
                ~gsub("\t", "", .x)))
# change columns in all dataset taken from "https://stackoverflow.com/questions/43456687/using-dplyr-gsub-on-many-columns"
```

### 3. Create .tsv files

```{r}
data_folder = "../data"

if (!dir.exists(data_folder)) {
  dir.create(data_folder)
}

metadata_df |> 
  write_tsv(file = str_c(data_folder, "/01_data_metadata.tsv"))

dictionary_df |> 
  write_tsv(file = str_c(data_folder, "/01_data_dictionary.tsv"))

otu_df |> 
  write_tsv(file = str_c(data_folder, "/01_data_otu.tsv"))
```
