==> R/03_augment.qmd <==
---
title: "Data augmentation"
author: Lucía de Lamadrid, Jorge Santiago Bajo, Eric Torres García, Elena Iriondo Delgado
format:
  html:
    embed-resources: true
editor: visual
---

### 1. Load packages

```{r}
#| message: false 
library("tidyverse") 
library("readr")
```

### 2. Read data

```{r}
#| message: false
clean_df <- read_tsv('../data/02_metadata_long_filtered_label.tsv')
head(clean_df)
```

In this section, we'll be working with the OTUs taxonomy file, so let's load it as well.

Note: The file `../data/01_data_otu.tsv` is misformatted. It is a TSV file but uses commas as separators. Therefore, we need to use `read.table` and explicitly specify the separator as a comma (`,`).

```{r}
otu_df_original <- read.table('../data/01_data_otu.tsv', 
                              header = TRUE, 
                              sep = ",")
head(otu_df_original)
```

If we do a deeper analysis of this dataset, we can actually check it is very ill-defined, if not directly erroneous. An example can be seen on OTU94, which presents several problems:

```{r}
otu_df_original |> 
  filter(OTU.ID == "OTU94")
```

The information from this OTU is the following:

-   Kingdom: Bacteria

-   Phylum: Actinobacteria

-   Class: Actinobacteria

-   Order: Actinobacteridae

Here comes the first problem! Actinobacteridae is actually a subclass, which is no longer in use (see <https://lpsn.dsmz.de/subclass/actinobacteridae>).

-   Family: Actinomycetales

Actinomycetales is actually an order, not a family.

-   Genus: Frankineae

Frankineae is an order, not a genus, and it is different from the previous order mentioned.

Additionally, there are two extra columns with no identification of taxonomy: Geodermatophilaceae and Blastococcus.

Such delirium of dataset is the result of complete lack of manual curation and care. As the authors from the paper we are inspired by use only the phylum and class to plot figures, we are going to choose the same criterium. These taxonomic ranks do not present the signalled problems with the other categories.

The taxonomical information of some OTUs is missing. For these cases, we are going to use the label "Unclassified":

```{r}
otu_df_modified <- otu_df_original |> 
  select(
    OTU.ID, 
    Phylum, 
    Class
  ) |> 
  mutate(Phylum = if_else(
    Phylum == "",
    "Unclassified",
    Phylum)
  ) |> 
  mutate(Class = if_else(
    Class == "", 
    "Unclassified", 
    Class)
  )

head(otu_df_modified)
```

After that, we perform a left join to include the taxonomical information into the data frame:

```{r}
clean_df_taxonomy <- clean_df |>  
  left_join(otu_df_modified, 
            join_by(OTU == OTU.ID)) |> 
  relocate(Phylum, Class, .after = OTU) 

head(clean_df_taxonomy)
```

```{r}
write_tsv(clean_df_taxonomy, 
            "../data/03_metadata_long_filtered_taxonomy.tsv")
```

Other analyses like PCA require a wide-format dataset but phylum based. To achieve this, we must first aggregate the data to ensure that each sample has a single value for every phylum before pivoting the data into a wide format:

```{r}
# Aggregate relative abundances by phylum
phylum_aggregated_data <- clean_df_taxonomy |>
  group_by(SampleID, Phylum, Diet, Donor) |> 
  summarize(rel_abundance = sum(rel_abundance), .groups = "drop")

# Pivot wider
phylum_aggregated_wide <- phylum_aggregated_data |> 
  pivot_wider(names_from = Phylum, 
              values_from = rel_abundance)

# Check of the aggregated data
head(phylum_aggregated_wide)
```

```{r}
write_tsv(phylum_aggregated_wide, 
          "../data/03_metadata_wide_phylum.tsv")
```

```{r}
write_tsv(otu_df_modified, 
          "../data/03_data_otu_modified.tsv")
```
