==> R/04_describe.qmd <==
---
title: "04_describe"
author: Eric Torres, Lucia de Lamadrid, Konstantina Gkopi, Elena Iriondo and Jorge Santiago
format:
  html:
    embed-resources: true
editor: visual
---

## Data visualization and description

### Load packages

```{r}
#| message: false 
library("tidyverse")
library("ggthemes")
```

### Read data

```{r}
#| message: false 
clean_df <- read_tsv('../data/03_metadata_long_filtered_taxonomy.tsv')
# head(clean_df)
```

### Microbiota composition differences due to sex

```{r}
clean_df |> 
  filter(Donor == "Fresh") |> # Filter for Donor == Fresh
  group_by(Sex, Phylum) |> # Group by Sex
  summarise(mean_rel_abund = mean(rel_abundance), 
            .groups = "drop") |> # Calculate mean relative abundances
  group_by(Sex) |> 
  mutate(mean_rel_abund = mean_rel_abund / sum(mean_rel_abund)) |> # Normalize mean abundances so that the total for each sex
  ggplot(aes(x = Sex,
             y = mean_rel_abund,
             fill = Phylum)) + 
  geom_bar(stat = "identity",
           position = "stack") +
  geom_hline(yintercept = 0) +
  labs(
    title = "Relative abundance of phyla by sex",
    x = "Sex",
    y = "Relative abundance",
    fill = "Phylum"
  ) +
  scale_fill_colorblind() +
  theme_minimal() +
  theme(legend.position = "bottom",
        legend.text = element_text(size = 10),
        axis.text.x = element_text(size = 10, angle = 45, hjust = 1),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 15), 
        axis.title.y = element_text(size = 15),
        plot.title = element_text(hjust = 0.5, size = 17))

```

As seen in the plot above there are no data for female mice in the first generation, therefore we cannot draw any conclusion. We wanted to use `Donor == "Fresh"` because when examining the differences due to `Sex`, the other variables like the transplantation or diet shouldn't be considered. If we take a look to the composition of the samples:

```{r}
clean_df |> 
  group_by(Donor, Sex) |> 
  summarise(count = n())
```

There are only female mice in the humanized mouse with LFPP diet graft, and the differences in sample size are huge in comparison to male individuals. Thus, this analysis cannot be done.

### Microbiota composition in terms of phyla and class in different sources and diet types

```{r}
plot1 <- clean_df |> 
  filter(Diet %in% c("LFPP", "Western")) |> 
  filter(Source != "Stomach") |> # Filter out the stomach samples as there are none in LFPP
  group_by(Source, Diet) |> 
  mutate(norm_rel_abundance = rel_abundance / sum(rel_abundance)) |> # Normalize abundance within samples
  group_by(Source, Diet, Phylum) |> 
  summarise(mean_rel_abund = mean(norm_rel_abundance), 
            .groups = "drop") |> # Calculate the mean relative abundances
  group_by(Source, Diet) |> 
  mutate(mean_rel_abund = mean_rel_abund / sum(mean_rel_abund)) |> # Normalize mean abundances so that the total for each one is 1
  ggplot(aes(x = Source,
             y = mean_rel_abund,
             fill = Phylum)) + 
  geom_bar(stat = "identity",
           position = "stack") +
  geom_hline(yintercept = 0) +
  facet_wrap(~ Diet) + 
  labs(
    title = "Relative abundance of phyla per source and diet type",
    x = "Source type",
    y = "Relative Abundance",
    fill = "Phyla"
  ) +
  scale_fill_colorblind() +
  theme_minimal() +
  theme(legend.position = "bottom", # change when data augmentation has been done
        legend.text = element_text(size = 10),
        axis.text.x = element_text(size = 10, angle = 45, hjust = 1),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 15), 
        axis.title.y = element_text(size = 15),
        plot.title = element_text(hjust = 0.5, size = 17),
        strip.text = element_text(size = 14))

ggsave("../results/rel_abund_phyla_source_diet.png", 
       plot = plot1, 
       width = 8, 
       height = 6, 
       units = "in", 
       bg = "white")

plot1
```

This plots represents the differences in microbiome phyla (in terms of relative abundance) found among different parts of the gut and feces in two different diets: LFFP and Western diet. Simply by observing the figure we can obtain valuable insights.

Firstly, we can find differences in phyla's relative abundance among the different parts of the mice's gut. However, the differences are noteworthy between diet types, showing a lower abundance of *Proteobacteria* and higher *Verrucomicrobia* in Western diet compared to the LFPP diet.

Another noteworthy observation is that in mice following an LFFP diet the *Firmicutes* abundance is more or less constant along the different sections of the gut, whereas in the Western diet there is a lower proportion in Cecum samples than in the SI samples.

These differences in relative abundance indicate a modulation of the microbiome by the diet type, and of course, it could have consequences on health. It would be interesting to analyse how this differential abundance could impact for example obesity risk, and also find out if there are some parts of the gut that have greater relevance in determining that risk.

### Microbiota composition of mice with different diet and donor combination (relative abundance of phyla)

In order to compare the general composition of the microbiota in the mice with different donor and diet, we may represent the relative abundance of the most prevalent phyla.

```{r}
plot2 <- clean_df |> 
  filter(Diet %in% c("LFPP", "Western") & Donor %in% c("Fresh", "HMouseLFPP", "HMouseWestern") & Source !="Feces") |> 
  group_by(Donor, Diet) |> 
  mutate(norm_rel_abundance = rel_abundance / sum(rel_abundance)) |> # Normalize abundance within samples
  group_by(Donor, Diet, Phylum) |> 
  summarise(mean_rel_abund = mean(norm_rel_abundance), 
            .groups = "drop") |> # Calculate the mean relative abundances
  group_by(Donor, Diet) |> 
  mutate(mean_rel_abund = mean_rel_abund / sum(mean_rel_abund)) |> # Normalize mean abundances so that the total for each one is 1
  ggplot(aes(x = Donor,
             y = mean_rel_abund,
             fill = Phylum)) + 
  geom_bar(stat = "identity",
           position = "stack") +
  geom_hline(yintercept = 0) +
  facet_wrap(~ Diet) + 
  labs(
    title = "Relative abundance of phyla per donor and diet type",
    x = "Donor type",
    y = "Relative Abundance",
    fill = "Phylum"
  ) +
  scale_fill_colorblind() +
  theme_minimal() +
  theme(legend.position = "bottom",
        legend.text = element_text(size = 10),
        axis.text.x = element_text(size = 10, angle = 45, hjust = 1),
        axis.text.y = element_text(size = 14),
        axis.title.x = element_text(size = 15), 
        axis.title.y = element_text(size = 15),
        plot.title = element_text(hjust = 0.5, size=17),
        strip.text = element_text(size = 14))

ggsave("../results/rel_abund_diet_donor_comb.png", 
       plot = plot2, 
       width = 8, 
       height = 6, 
       units = "in", 
       bg = "white")

plot2
```

This plot is very informative, as it allows us to explore several of the different experiments carried out by the original authors of the dataset:

-   Overall, we can see that the Western diet leads to a lower proportion of Bacteroidetes and unclassified phyla, and a higher one of Verrucomicrobia*.*

-   When focusing on the first generation of humanised mice ("Fresh" donor), we can see that diet greatly affects the microbiota composition. The LF-PP diet leads to a much higher proportion of Bacteroidetes and a smaller one of Actinobacteria. Furthermore, there are less bacteria belonging to unclassified phyla in the Western diet, and more Verrucomicrobia*.*

-   Regarding the second generation of humanised mice ("Hmouse..." donor), we can observe that the bars with the same donor are more similar to each other ("HmouseLFPP" versus "HmouseWestern") than those with the same diet ("LFPP" vs "Western"). This indicates that the diet that their "donor" followed may be more determinant in the composition of their gut microbiota than the diet the mice follow themselves.

-   Though the diet factor has the most obvious effect on the first generation of humanised mice, the diet followed also visibly influenced the microbiota of the second generation of humanised mice. For instance, mice whose donors were on the Western diet ("HmouseWestern" donor) presented starkly different proportions of Verrucomicrobia and Bacteroidetes depending on their current diet.
