==> R/05_analysis_1.qmd <==
---
title: "05_analysis_1"
author: Eric Torres, Lucía de Lamadrid, Elena Iriondo, Konstantina Gkopi and Jorge Santiago
format:
  html:
    embed-resources: true
editor: visual
---

### 1. Load packages

```{r}
#| message: false 
library("tidyverse")
library("ggrepel")
library("ggtext")
library("broom") 
library("readr")
```

### 2. Read data

```{r}
#| message: false 
clean_df <- read_tsv(file = '../data/03_metadata_long_filtered_taxonomy.tsv') 
otu_df_modified <- read_tsv(file = '../data/03_data_otu_modified.tsv')
```

### 3. Analysis

#### Firmicutes-Bacteroidetes Ratio

The authors Turnbaugh *et al.* attempt to induce obesity in the mice by providing a Western diet. Both weight and body fat were quantified to measure this condition. However, this data was not provided, so we must resort to other indirect measures of obesity. One of these is the Firmicutes-Bacteroidetes ratio, a well-known biomarker for obesity (<https://doi.org/10.1038/4441022a>).

As an initial approach to see the "obesity-inducing" effects of this diet on the humanized mice, we will calculate the Firmicutes-Bacteroidetes ratio in first generation humanized mice (donor is "Fresh").

```{r}
ratio_data <- clean_df |> 
  filter(Diet %in% c("LFPP", "Western") & Donor == "Fresh") |>
  group_by(Diet) |> 
  summarise(Firmicutes = sum(rel_abundance[Phylum == "Firmicutes"]),
            Bacteroidetes = sum(rel_abundance[Phylum == "Bacteroidetes"])) |> 
  mutate(Ratio = Firmicutes / Bacteroidetes)

ratio_data
```

As expected, the ratio is bigger in the Western diet of first generation humanized mice (less Bacteroidetes and more Firmicutes proportionately).

To investigate this phenomenon more rigorously, we will develop and analyse models. Specifically, we want to test if there are differences in the relative abundance of the OTUs belonging to Firmicutes or Bacteroidetes depending on the diets LFPP or Western.

First, we will filter for the two types of diets and the two phyla we will be working with:

```{r}
diet_phylum <- clean_df |> 
  filter(Phylum == "Firmicutes" | Phylum == "Bacteroidetes") |>  
  filter(Diet == "LFPP" | Diet == "Western") |>  
  mutate(Diet = case_when(Diet == "LFPP" ~ 0, 
                         Diet == "Western" ~ 1))
```

Then, we create a nested data frame:

```{r}
diet_phylum_nested <- diet_phylum |> 
  group_by(OTU) |> 
  nest() |> 
  ungroup()
```

To identify which OTUs are significantly associated with diet, a linear model is fitted for each OTU to assess the relationship between relative abundance and diet.

```{r}
diet_phylum_nested <- diet_phylum_nested |> 
  group_by(OTU) |>  
  mutate(model_object = map(.x = data, 
                             .f = ~lm(formula = rel_abundance ~ Diet, 
                                      data = .x)))
```

The model results are tidied to extract estimates, p-values and confidence intervals. Afterward, p-values are adjusted for multiple testing, and OTUs with significant associations to diet are identified based on a threshold (q-value \< 0.05).

```{r}
otus_estimate <- diet_phylum_nested |> 
  mutate(model_object_tidy = map(.x = model_object, 
                                 .f = ~tidy(.x, conf.int = TRUE, conf.level = 0.95))
  ) |>  
  unnest(model_object_tidy) |> 
  filter(term == "Diet") |>
  select (OTU, 
          p.value, 
          estimate, 
          conf.low, 
          conf.high) |> 
  ungroup() |> 
  mutate(
    # Adjust p-values
    q.value = p.adjust(p.value), 
    is_significant = case_when( 
      q.value >= 0.05 ~ "no",
      q.value < 0.05 ~ "yes"
      )
  )
```

We will select the OTUs whose relative abundance significantly changes between the two dietary groups:

```{r}
otus_significant <- otus_estimate |>  
  left_join(otu_df_modified, 
            join_by(OTU == OTU.ID)) |> 
  relocate(Phylum, Class, .after = OTU) |> 
  filter(is_significant == "yes") 
```

We can see that a large proportion of OTUs have a relative abundance which is significantly different between both diet types (153 out of 355).

We are going to represent the estimates (95% confidence intervals) of the 30 OTUs with the highest statistical significance (lower q-value):

```{r}
otus_western_diet <- otus_significant |> 
  arrange(q.value) |> 
  slice_head(n = 30) |> 
  # Plot reordering OTUs by their effect estimates
  ggplot(
    aes(x = estimate, 
        y = fct_reorder(OTU, estimate), 
        color = Phylum)
    ) +
    geom_point(size = 2) +
    # Add horizontal error bars for confidence intervals
    geom_errorbarh(
      aes(xmax = conf.high, 
          xmin = conf.low)
    ) +
    # Add a vertical line at 0 for reference
    geom_vline(
      xintercept = 0, 
      linetype = "dashed", 
      color = "gray"
    ) +
    labs(
      title = "Top 30 OTUs with significant different \n relative abundances according to diet",
      subtitle = "Colored by phylum",
      x = "Estimates (95% confidence intervals)",
      y = "OTU"
    ) +
    theme_minimal() +
    theme(
      plot.title = element_text(hjust = 0.5, size = 14),
      plot.subtitle = element_text(hjust = 0.5, size = 10)
    )

ggsave("../results/otus_associated_western_diet_model.png", 
       plot = otus_western_diet, 
       width = 8, 
       height = 6, 
       units = "in", 
       bg = "white")

otus_western_diet
```

Most OTUs with positive estimates (higher abundance in the Western Diet, 1) belong to Firmicutes, while some Bacteroidetes are negatively associated. This observation aligns to the results from the ration.

By focusing on a specific OTU that showed significant variation in the abundance between the two diets (e.g. OTU5691), we confirmed that there is a clear shift in its abundance, indicating that this metric is influenced by the dietary conditions.

```{r}
diet_phylum |>  
  filter(OTU == "OTU5691") |> 
  group_by(Diet) |>          
  summarize(total_rel_abundance = mean(rel_abundance))
```

Moreover, counting the number of OTUs associated with each diet and phylum helps identify trends, such as a particular phylum showing a stronger association with one diet compared to the other.

```{r}
vector_otus <- otus_significant |> 
  pull(OTU)

result <- diet_phylum |> 
  filter(OTU %in% vector_otus) |>  
  group_by(OTU, Diet) |> 
  summarize(total_rel_abundance = mean(rel_abundance), 
            .groups = "drop") |> 
  group_by(OTU) |> 
  # Find which diet presents higer average abundance for each OTU 
  mutate(higher_in = if_else(
    total_rel_abundance[Diet == 0] > total_rel_abundance[Diet == 1], 
    "LFPP Diet", 
    "Western Diet")) |> 
  distinct(OTU, higher_in)

# Add taxonomic information
result <- result |>  
  left_join(otu_df_modified, 
            join_by(OTU == OTU.ID)) |> 
  select(OTU, 
         Phylum, 
         higher_in)
```

We represent a contigency table with the results and then show a bar plot:

```{r}
contingency_table <- result |> 
  group_by(higher_in, Phylum) |> 
  summarize(count = n(), 
            .groups = "drop") |> 
  pivot_wider(names_from = Phylum, 
              values_from = count)

contingency_table
```

```{r}
plot_data <- result |> 
  group_by(Phylum, higher_in) |> 
  summarize(count = n(), 
            .groups = "drop")

plot_otu_abundance_diet <- plot_data |>
  ggplot(
    aes(x = Phylum, 
        y = count,
        fill = higher_in)
    ) +
    geom_bar(stat = "identity", position = "dodge") + 
    labs(
      title = "Number of OTUs with higher relative abundance by diet",
      subtitle = "Only OTUs with significant different abundance between diets were considered",
      x = "Phylum",
      y = "Number of OTUs",
      fill = "Higher in"
    ) +
    theme_minimal() +
    theme(plot.title = element_text(hjust = 0.5, size = 14),
          plot.subtitle = element_text(hjust = 0.5, size = 10))

ggsave("../results/abundance_otu_phylum_diet.png", 
       plot = plot_otu_abundance_diet, 
       width = 8, 
       height = 6, 
       units = "in", 
       bg = "white")

plot_otu_abundance_diet
```

A substantial number of OTUs belonging to the phylum Firmicutes exhibit higher relative abundance in the Western diet than in LFPP, indicating that Firmicutes are likely more influenced by the characteristics of diets high in fats and sugars. The more balanced distribution between Bacteroidetes and Firmicutes in the LFPP diet suggests a healthier microbiota composition, often associated with diets rich in fiber and low in processed foods.
