==> R/07_analysis_3.qmd <==
---
title: "07_analysis_3"
author: "Eric Torres, Lucía de Lamadrid, Elena Iriondo, Konstantina Gkopi and Jorge Santiago"
format:
  html:
    embed-resources: true
editor: visual
---

Load packages

```{r}
#| message: false 
library("tidyverse")
```

Read data

```{r}
#| message: false 
wide_df <- read_tsv('../data/02_metadata_wide_filtered_label.tsv')
#head(wide_df)
```

### Biodiversity of humanised mice microbiota with different diet

We can use diversity indices in order to quantify biodiversity. As we know the relative abundance of each OTU, we can use the Shannon diversity index. Based on Claude Shannon's formula for entropy, this popular ecological metric takes into account the number of species living in a habitat (richness) and their relative abundance (evenness).

The formula we will use is:

$$ H' = -\sum_{i=1}^R p_i \ln p_i \\ p_i \text{ is the relative abundance of OTU}_i\\ R \text{ is the total number of OTUs} $$

We will compare this index for first generation humanised mice following each diet. For this purpose, we will use the wide data, in which the rows include all OTU counts for each sample.

We calculate the Shannon index for each row and save it in a new dataframe.

```{r}
shannon <- wide_df |> 
  filter(Donor == "Fresh" & Source != "Feces" & Diet %in% c("Western", "LFPP")) |>#first generation ("fresh" donor), only take into account "final" snapshots of the microbiota (no intermediate fecal samples), and we make sure we are only comparing the two diets 
  group_by(Diet) |> 
  select(starts_with("OTU")) |> #we only need the OTU columns to calculate the index 
  mutate(across(everything(), ~ (-1)*.x * log(.x))) |> #we transform each OTU relative abundance column (p_i) into p_i*log(p_i) 
  rowwise() |> #we apply the following operations row by row
  mutate(
    sh = sum(across(everything()), na.rm = TRUE), #we add the products to calculate the shannon_index
    .keep = "unused") |> #we only keep the new shannon index column and the diet column
  ungroup()
head(shannon)
```

We now carry out a t test in order to compare the shannon index of both populations (western versus LF-PP diet). As the sample size is large, we do not test for normality previously.

```{r}
#| warning: false 
t_test_result <- shannon |> 
  summarise(
    t_test = list(t.test(sh ~ Diet, data = cur_data()))) |> 
  pull(t_test)
t_test_result
```

The result of the *t* test is that the means of the Shannon index for each diet are (very) significantly different. We can now represent this result in a barplot, showing the standard deviation as error.

```{r}
shannon_plot <-shannon |> group_by(Diet) |>
  summarise(shannon_index= mean(sh),
    standard_dev = sd(sh)) |>
  ggplot(aes(x = Diet,
             y = shannon_index)) + 
  geom_bar(stat = "identity",
           position = "stack", aes(fill = Diet, alpha=0.5), width = 0.6) +
  geom_errorbar(aes(ymin = shannon_index - standard_dev, ymax = shannon_index + standard_dev), 
                width = 0.2) +  # Error bar width
  labs(
    title = "Biodiversity in microbiota",
    x = "Diet",
    y = "Shannon diversity index"
  ) +
  theme_minimal() +
  theme(
    legend.position = "none",  # Remove legend
    axis.text.x = element_text(size = 14, hjust = 0.5),
    axis.text.y = element_text(size = 14),
    axis.title.x = element_text(size = 15), 
    axis.title.y = element_text(size = 15),
    plot.title = element_text(hjust = 0.5, size = 17)
  )

ggsave("../results/shannon.png", 
       plot = shannon_plot, 
       width = 8, 
       height = 6, 
       units = "in", 
       bg = "white")

shannon_plot
```

Interestingly, the Western diet leads to a more biodiverse microbiota than the "healthier" LF-PP diet in the humanised mice. This is in line with the results of representing the microbiota composition in function of phylum, in which the different type of phyla were more evenly distributed in the Western diet.
